+++ src/history.settings.ts
@@ -11,7 +11,9 @@ const enum EHistoryEnabled {
 export interface IHistorySettings {
     folder: vscode.Uri;
     daysLimit: number;
+    saveDelay: number;
     maxDisplay: number;
+    dateLocale: string;
     exclude: string[];
     enabled: boolean;
     historyPath: string;
@@ -125,7 +127,9 @@ export class HistorySettings {
         return {
             folder: workspacefolder,
             daysLimit: <number>config.get('daysLimit') || 30,
+            saveDelay: <number>config.get('saveDelay') || 0,
             maxDisplay: <number>config.get('maxDisplay') || 10,
+            dateLocale: <string>config.get('dateLocale') || undefined,
             exclude: <string[]>config.get('exclude') || ['**/.history/**','**/.vscode/**','**/node_modules/**','**/typings/**','**/out/**'],
             enabled: historyPath != null && historyPath !== '',
             historyPath: historyPath,